+++ src/zrb/builtin/project/add/fastapp/crud/_helper/__init__.py
@@ -1,0 +1,7 @@
+from .register_api import register_api
+from .register_permission import register_permission
+from .register_rpc import register_rpc
+
+assert register_api
+assert register_permission
+assert register_rpc
